FileUtil: add a sanity check to prevent silly bugs
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/freenet@15967 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/support/io/FileUtil.java
@@ -147,7 +147,8 @@ final public class FileUtil {
         public static boolean renameTo(File orig, File dest) {
             // Try an atomic rename
             // Shall we prevent symlink-race-conditions here ?
-            
+            if(orig.equals(dest))
+                throw new IllegalArgumentException("Huh? the two file descriptors are the same!");
             if (!orig.renameTo(dest)) {
                 // Not supported on some systems (Windows)
                 if (!dest.delete()) {
